+++ R/mage.R
@@ -33,9 +33,13 @@
 #' data(example_data_5_subject)
 #' mage(example_data_5_subject, version = 'ma')
 
-mage <- function(data, version = c('ma', 'naive'), sd_multiplier = 1,
-                 short_ma = 5, long_ma = 32, type = c('auto', 'plus', 'minus'),
-                 plot = FALSE, dt0 = NULL, inter_gap = 45, tz = "",
+mage <- function(data,
+                 version = c('ma', 'naive'),
+                 sd_multiplier = 1,
+                 short_ma = 5, long_ma = 32,
+                 return_type = c('num', 'df'),
+                 direction = c('service', 'avg', 'max', 'plus', 'minus'),
+                 plot = FALSE, dt0 = NULL, inter_gap = 45, max_gap=180, tz = "",
                  title = NA, xlab = NA, ylab = NA, show_ma = FALSE) {
 
   # Match version
@@ -46,25 +50,28 @@ mage <- function(data, version = c('ma', 'naive'), sd_multiplier = 1,
     return(mage_sd(data, sd_multiplier = sd_multiplier))
   }
 
-  return(mage_ma(data, short_ma = short_ma, long_ma = long_ma, type = type,
-                 plot = plot, dt0 = dt0, inter_gap = inter_gap, tz = tz,
+  return(mage_ma(data, short_ma = short_ma, long_ma = long_ma, return_type=return_type, direction=direction,
+                 plot = plot, dt0 = dt0, inter_gap = inter_gap, max_gap = max_gap, tz = tz,
                  title = title, xlab = xlab, ylab = ylab, show_ma = show_ma))
 }
 
-mage_ma <- function(data, short_ma = 5, long_ma = 32, type = c('auto', 'plus', 'minus'),
-                    plot = FALSE, dt0 = NULL, inter_gap = 45, tz = "",
+mage_ma <- function(data,
+                    short_ma = 5, long_ma = 32,
+                    return_type = c('num', 'df'),
+                    direction = c('service', 'avg', 'max', 'plus', 'minus'),
+                    plot = FALSE, dt0 = NULL, inter_gap = 45, max_gap = 180, tz = "",
                     title = NA, xlab = NA, ylab = NA, show_ma = FALSE) {
   id = . = MAGE = NULL
   rm(list = c("id", ".", "MAGE"))
 
   data = check_data_columns(data)
-  is_vector = attr(data, "is_vector") # TODO: is this check really necessary? when does this return true?
+  is_vector = attr(data, "is_vector")
 
   out <- data %>%
     dplyr::filter(!is.na(gl)) %>%
     dplyr::group_by(id) %>%
-    dplyr::do(MAGE = mage_ma_single(., short_ma = short_ma, long_ma = long_ma, type = type,
-                                    plot = plot, dt0 = dt0, inter_gap = inter_gap, tz = tz,
+    dplyr::do(MAGE = mage_ma_single(., short_ma = short_ma, long_ma = long_ma, return_type=return_type, direction=direction,
+                                    plot = plot, dt0 = dt0, inter_gap = inter_gap, max_gap = max_gap, tz = tz,
                                     title = title, xlab = xlab, ylab = ylab, show_ma = show_ma))
 
   # Check if a ggplot or number in list is returned - convert the latter to a number